opencl: cast mxfp4 scale to half in gemm_moe_mxfp4_f32_ns (Adreno 850 compiler)
mxfp4_to_fp16_packed8(...) * s multiplies a half8 vector by a float
scalar. That mixed-type vector*scalar multiply crashes the Adreno 850
shader compiler (driver art.api37) during clBuildProgram -- the process
dies with no CL_BUILD_PROGRAM_FAILURE and no build log -- which aborts
backend init whenever this kernel is compiled (it is built eagerly, even
for non-MXFP4 models).

Cast the scale to half explicitly (* (half)s). OpenCL's usual arithmetic
conversions already convert the scalar to the vector element type, so the
result is byte-identical on every other driver -- no behavior or
performance change (verified on Adreno 840: gpt-oss-20B MXFP4 prefill
unchanged, MUL_MAT_ID mxfp4 test-backend-ops OK).

+++ ggml/src/ggml-opencl/kernels/gemm_moe_mxfp4_f32_ns.cl
@@ -273,9 +273,12 @@ kernel void kernel_gemm_moe_mxfp4_f32_ns(
         shared_b[b_local_offset.x] = bx8_f16.lo;
         shared_b[b_local_offset.y] = bx8_f16.hi;
 
-        // Dequantization
-        reg_a.lo = mxfp4_to_fp16_packed8(as_ushort2(mxfp4x16.lo)) * s;
-        reg_a.hi = mxfp4_to_fp16_packed8(as_ushort2(mxfp4x16.hi)) * s;
+        // Dequantization. Cast the scale to half explicitly: a bare half8*float
+        // (vector*scalar mixed-type) multiply crashes the Adreno 850 (art.api37)
+        // shader compiler. OpenCL converts the scalar to the vector element type
+        // anyway, so this is byte-identical on every other driver.
+        reg_a.lo = mxfp4_to_fp16_packed8(as_ushort2(mxfp4x16.lo)) * (half)s;
+        reg_a.hi = mxfp4_to_fp16_packed8(as_ushort2(mxfp4x16.hi)) * (half)s;
 
         sub_group_barrier(CLK_LOCAL_MEM_FENCE);
 
@@ -303,9 +306,12 @@ kernel void kernel_gemm_moe_mxfp4_f32_ns(
         shared_b[b_local_offset.x] = bx8_f16.lo;
         shared_b[b_local_offset.y] = bx8_f16.hi;
 
-        // Dequantization
-        reg_a.lo = mxfp4_to_fp16_packed8(as_ushort2(mxfp4x16.lo)) * s;
-        reg_a.hi = mxfp4_to_fp16_packed8(as_ushort2(mxfp4x16.hi)) * s;
+        // Dequantization. Cast the scale to half explicitly: a bare half8*float
+        // (vector*scalar mixed-type) multiply crashes the Adreno 850 (art.api37)
+        // shader compiler. OpenCL converts the scalar to the vector element type
+        // anyway, so this is byte-identical on every other driver.
+        reg_a.lo = mxfp4_to_fp16_packed8(as_ushort2(mxfp4x16.lo)) * (half)s;
+        reg_a.hi = mxfp4_to_fp16_packed8(as_ushort2(mxfp4x16.hi)) * (half)s;
 
         sub_group_barrier(CLK_LOCAL_MEM_FENCE);
